refactor: move `Allowance.json` into `src/logic/contracts/artifacts` directory
- also reorganized imports in affected files
