+++ src/logic/contracts/safeContracts.ts
@@ -1,19 +1,21 @@
-import { SPENDING_LIMIT_MODULE_ADDRESS } from 'src/utils/constants'
-import { AbiItem } from 'web3-utils'
-import contract from 'truffle-contract'
-import Web3 from 'web3'
-import ProxyFactorySol from '@gnosis.pm/safe-contracts/build/contracts/GnosisSafeProxyFactory.json'
 import GnosisSafeSol from '@gnosis.pm/safe-contracts/build/contracts/GnosisSafe.json'
 import SafeProxy from '@gnosis.pm/safe-contracts/build/contracts/GnosisSafeProxy.json'
-import SpendingLimitModule from 'src/utils/AllowanceModule.json'
-import { ensureOnce } from 'src/utils/singleton'
+import ProxyFactorySol from '@gnosis.pm/safe-contracts/build/contracts/GnosisSafeProxyFactory.json'
 import memoize from 'lodash.memoize'
-import { getWeb3, getNetworkIdFrom } from 'src/logic/wallets/getWeb3'
-import { calculateGasOf, calculateGasPrice } from 'src/logic/wallets/ethTransactions'
-import { ZERO_ADDRESS } from 'src/logic/wallets/ethAddresses'
+import contract from 'truffle-contract'
+import Web3 from 'web3'
+import { AbiItem } from 'web3-utils'
+
 import { isProxyCode } from 'src/logic/contracts/historicProxyCode'
-import { GnosisSafeProxyFactory } from 'src/types/contracts/GnosisSafeProxyFactory.d';
+import { ZERO_ADDRESS } from 'src/logic/wallets/ethAddresses'
+import { calculateGasOf, calculateGasPrice } from 'src/logic/wallets/ethTransactions'
+import { getNetworkIdFrom, getWeb3 } from 'src/logic/wallets/getWeb3'
 import { GnosisSafe } from 'src/types/contracts/GnosisSafe.d'
+import { GnosisSafeProxyFactory } from 'src/types/contracts/GnosisSafeProxyFactory.d'
+import { SPENDING_LIMIT_MODULE_ADDRESS } from 'src/utils/constants'
+import { ensureOnce } from 'src/utils/singleton'
+
+import SpendingLimitModule from './artifacts/AllowanceModule.json'
 
 export const SENTINEL_ADDRESS = '0x0000000000000000000000000000000000000001'
 export const MULTI_SEND_ADDRESS = '0x8d29be29923b68abfdd21e541b9374737b49cdad'
