+++ src/routes/safe/components/Settings/SpendingLimit/RemoveLimitModal.tsx
@@ -5,6 +5,7 @@ import { useDispatch, useSelector } from 'react-redux'
 
 import Block from 'src/components/layout/Block'
 import Col from 'src/components/layout/Col'
+import SpendingLimitModule from 'src/logic/contracts/artifacts/AllowanceModule.json'
 import createTransaction from 'src/logic/safe/store/actions/createTransaction'
 import { safeParamAddressFromStateSelector } from 'src/logic/safe/store/selectors'
 import { TX_NOTIFICATION_TYPES } from 'src/logic/safe/transactions'
@@ -12,12 +13,11 @@ import { ETH_ADDRESS } from 'src/logic/tokens/utils/tokenHelpers'
 import { ZERO_ADDRESS } from 'src/logic/wallets/ethAddresses'
 import { getWeb3 } from 'src/logic/wallets/getWeb3'
 import useToken from 'src/routes/safe/components/Settings/SpendingLimit/hooks/useToken'
-import SpendingLimitModule from 'src/utils/AllowanceModule.json'
 import { SPENDING_LIMIT_MODULE_ADDRESS } from 'src/utils/constants'
 
 import { SpendingLimitTable } from './dataFetcher'
-import { AddressInfo, TokenInfo, ResetTimeInfo } from './InfoDisplay'
 import { RESET_TIME_OPTIONS } from './FormFields/ResetTime'
+import { AddressInfo, ResetTimeInfo, TokenInfo } from './InfoDisplay'
 import Modal from './Modal'
 import { useStyles } from './style'
 import { fromTokenUnit } from './utils'
